connected react with Redux
import Provider from react-redux and wrap it with App.
for the provider give the store as props.
in the component in wich you want to use redux state, import connect from react-redux.

while exporting the component call connect with the sliced state as a parameter.

eg. const slicedState = (state) => { return { ctr: state.counter } }

export default connect( slicedState ) (Component)

the state is from the reducer . because you connected it with the redux using react-redux.

now you can use the state in the component. It is given as props.