Added dispatch to React
Same as we recieve state  as props from reducer, we recieve dispatch from the reducer.

create a function getting dispatch as argument, it then returns an object containing types of functions to be executed . those functions are dispatch functions that has an argument which is an object specifying the type, and the value to perforn certain actions.

pass the function to the connect function as an argument.

now the actions to be performed are recieved as props for the component.
we can use it by calling it onclick={  this.props.onIncreament  }