Using our own middle ware.
A middleware is a function that takes store ( the new store from the reducer ) as an argument , returns a function which takes the next ( used to connect middleware and reducer using the action as parameter ), it then returns a function with argument action ( the action we specified ) , it then returns the const which stores the result of the execution of the method next.

now you can acces
      1. the action you perform also can alter the action in middleware and send it to reducer.
      2. the next function ( takes old action or altered action as a parameter ) ,store the result in a const and return it in the last function which takes action as parameter.
      3.  the store ( which gives new altered state )

Now all we have to do is pass the middleware to the create store function as a second argument.
For that use a hoc provided by redux , applyMiddleware which is a function takes middlewares as arguments.