Used redux thunk.
It is used to perform async actions .

It is a middleware so import it from redux-thunk and pass it as an argument for applyMiddleware.

now actually we arre going to dispatch the action again while irt is being dispatched.

For that go to actions where a function returns the argument for the dispatch ,  in tha tfunction it returns a function which gets dispatch as an argument. Now inside the fiunction run your async codes eg. setTimeout. inside the async code execute the dispatch( { type: ... , payLoad: { ... } ) function with the arguments it want to recieve.